itemsales commit

+++ Models/DepartmentDto.cs
@@ -4,7 +4,7 @@ namespace Sharp.Models
     {
         public int Id{get;set;}
         public string Department{get;set;}="";
-        public long Qty{get;set;}
+        public decimal Qty{get;set;}
         public decimal Amount{get;set;}
         public int SrvNo {get;set;}
         public int Till{get;set;}
